pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orbital-mission-compiler"
version = "0.1.0"
description = "Mission-plan-aware workflow compilation scaffold inspired by ORCHIDE"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML==6.0.2",
  "pydantic==2.12.5",
  "hera-workflows==6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest==9.0.2",
  "ruff==0.15.8",
]
mcp = [
  "fastmcp==3.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
